поправил вывод комментариев
