+++ src/components/movie-tabs/movie-tabs.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import MovieDetails from '../movie-details/movie-details';
 import MovieOverview from '../movie-overview/movie-overview';
 import MovieReview from '../movie-reviews/movie-reviews';
-import {MOVIES_PROP, REVIEW_PROP} from '../../utils/validate';
+import {MOVIES_PROP} from '../../utils/validate';
 
-const MovieTabs = ({film, reviews}) => {
+const MovieTabs = ({film}) => {
 
   const getActiveTab = () => {
     if (activeTab.details) {
@@ -25,7 +25,7 @@ const MovieTabs = ({film, reviews}) => {
         starring={film.starring}
       />;
     }
-    return <MovieReview reviews={reviews} />;
+    return <MovieReview filmID={film.id} />;
   };
 
   const [activeTab, setActiveTab] = useState({overview: true, details: false, review: false});
@@ -72,7 +72,6 @@ const MovieTabs = ({film, reviews}) => {
 
 MovieTabs.propTypes = {
   film: PropTypes.shape(MOVIES_PROP).isRequired,
-  reviews: PropTypes.arrayOf(PropTypes.shape(REVIEW_PROP))
 };
 
 export default MovieTabs;
