+++ src/components/movie/movie.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {Link} from 'react-router-dom';
 import {connect} from 'react-redux';
@@ -7,13 +7,11 @@ import MoviesList from '../movies-list/movies-list';
 import MovieTabs from '../movie-tabs/movie-tabs';
 import AddFavorite from '../add-favorite/add-favorite';
 import {AuthorizationStatuses, Url} from '../../consts';
-import {MOVIES_PROP, REVIEW_PROP} from '../../utils/validate';
+import {MOVIES_PROP} from '../../utils/validate';
 import {getSimmilarMoviesWithGenre} from '../../store/films/selectors';
 import {getAuthorizationStatus} from '../../store/auth/selectors';
-import {getReviews} from '../../store/comment/selectors';
-import {fetchComments} from '../../store/api-actions';
 
-const Movie = ({film, reviews, films, onPlayMovie, authorizationStatus, loadComments}) => {
+const Movie = ({film, films, onPlayMovie, authorizationStatus}) => {
   const {
     backgroundImage,
     name,
@@ -23,12 +21,6 @@ const Movie = ({film, reviews, films, onPlayMovie, authorizationStatus, loadComm
     id,
     isFavorite
   } = film;
-
-  useEffect(() => {
-    if (reviews[id] === undefined) {
-      loadComments(id);
-    }
-  }, [reviews]);
 
   return (
     <React.Fragment>
@@ -82,7 +74,6 @@ const Movie = ({film, reviews, films, onPlayMovie, authorizationStatus, loadComm
             <div className="movie-card__desc">
               <MovieTabs
                 film={film}
-                reviews={reviews[id]}
               />
             </div>
           </div>
@@ -115,23 +106,14 @@ const Movie = ({film, reviews, films, onPlayMovie, authorizationStatus, loadComm
 Movie.propTypes = {
   films: PropTypes.arrayOf(PropTypes.shape(MOVIES_PROP)).isRequired,
   film: PropTypes.shape(MOVIES_PROP).isRequired,
-  reviews: PropTypes.objectOf(PropTypes.arrayOf(PropTypes.shape(REVIEW_PROP))),
   onPlayMovie: PropTypes.func.isRequired,
-  loadComments: PropTypes.func.isRequired,
   authorizationStatus: PropTypes.string.isRequired
 };
 
 const mapStateToProps = (state) => ({
   films: getSimmilarMoviesWithGenre(state),
   authorizationStatus: getAuthorizationStatus(state),
-  reviews: getReviews(state)
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  loadComments(id) {
-    dispatch(fetchComments(id));
-  },
 });
 
 export {Movie};
-export default connect(mapStateToProps, mapDispatchToProps)(Movie);
+export default connect(mapStateToProps)(Movie);
